# Aula12_28Mai/serverAula12/apidoc.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Um título de API",
    "description": "Descrição da API"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": {
    "Bearer": []
  },
...
  "paths": {
    "/auth": {
      "post": {
        "tags": [
          "AUTHORIZATION"
        ],
        "summary": "Autoriza e gera token",
        "description": "Devolve um token válido",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "auth",
            "description": "Objeto que descreve um autorização",
            "schema": {
              "type": "object",
              "required": [
                "senha", "email"
              ],
              "properties": {
                "email": { "type": "string", "example": "ana@ana.com" },
                "senha": { "type": "string", "example": "12345" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação bem sucedida"
          },
          "400": {
            "description": "Operação inválida"
          }
        }
      }
    },
...
    "/usuarios": {
      "get": {
        "tags": [
          "USUARIOS"
        ],
        "summary": "Retorna usuarios",
        "description": "Retorna todos os usuarios existentes",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Operação bem sucedida"
          }
        }
      },
      "post": {
        "tags": [
          "USUARIOS"
        ],
        "summary": "Adiciona usuário",
        "description": "Adiciona um usuário na lista de usuários",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "usuario",
            "description": "Objeto que descreve um usuário",
            "schema": {
              "type": "object",
              "required": [
                "nome", "login", "senha", "email"
              ],
              "properties": {
                "nome": { "type": "string", "example": "Ana" },
                "login": { "type": "string", "example": "ana" },
                "senha": { "type": "string", "example": "12345" },
                "email": { "type": "string", "example": "ana@ana.com" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação Inserção bem sucedida"
          }
        }
      }
    }
  }
}